backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.136.0",
    "feedparser>=6.0.12",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
